ecom-microservices/recommendation-service/requirements.txt
werkzeug==2.0.3
flask==2.0.1
flask-cors==3.0.10
requests==2.31.0
python-dotenv==1.0.0
redis==5.0.1
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
scipy==1.11.3
marshmallow==3.20.1
gunicorn==21.2.0
pyarrow==14.0.0
cachetools==5.3.1
Flask-Limiter==2.8.0
pytest==7.4.2
pytest-cov==4.1.0 